formatting on save as command
