+++ src/format.ts
@@ -1,0 +1,52 @@
+/*---------------------------------------------------------
+ * Copyright (C) Microsoft Corporation. All rights reserved.
+ *--------------------------------------------------------*/
+
+'use strict';
+
+import vscode = require('vscode');
+import cp = require('child_process');
+import path = require('path');
+import os = require('os');
+import fs = require('fs');
+import {getBinPath} from './goPath'
+
+export class GoFormatter {
+
+	private formatCommand = "goreturns";
+
+
+	constructor() {
+		let formatTool = vscode.workspace.getConfiguration('go')['formatTool'];
+		if (formatTool) {
+			this.formatCommand = formatTool;
+		}
+	}
+
+	public provideDocumentFormattingEdits(document: vscode.TextDocument): Thenable<vscode.TextEdit[]> {
+		return new Promise((resolve, reject) => {
+			var filename = document.fileName;
+
+			var formatCommandBinPath = getBinPath(this.formatCommand);
+
+			cp.execFile(formatCommandBinPath, [filename], {}, (err, stdout, stderr) => {
+				try {
+					if (err && (<any>err).code == "ENOENT") {
+						vscode.window.showInformationMessage("The 'goreturns' command is not available.  Use 'go get -u sourcegraph.com/sqs/goreturns' to install.");
+						return resolve(null);
+					}
+					if (err) return reject("Cannot format due to syntax errors.");
+					var text = stdout.toString();
+					// TODO: Should use `-d` option to get a diff and then compute the
+					// specific edits instead of replace whole buffer
+					var lastLine = document.lineCount;
+					var lastLineLastCol = document.lineAt(lastLine - 1).range.end.character;
+					var range = new vscode.Range(0, 0, lastLine - 1, lastLineLastCol);
+					return resolve([new vscode.TextEdit(range, text)]);
+				} catch (e) {
+					reject(e);
+				}
+			});
+		});
+	}
+}
