+++ src/goFormat.ts
@@ -8,20 +8,12 @@ import vscode = require('vscode');
 import cp = require('child_process');
 import path = require('path');
 import {getBinPath} from './goPath'
+import { GoFormatter } from './format';
 
 export class GoDocumentFormattingEditProvider implements vscode.DocumentFormattingEditProvider {
 
-	private formatCommand = "goreturns";
-
 	// Not used?
 	public autoFormatTriggerCharacters: string[] = [';', '}', '\n'];
-
-	constructor() {
-		let formatTool = vscode.workspace.getConfiguration('go')['formatTool'];
-		if (formatTool) {
-			this.formatCommand = formatTool;
-		}
-	}
 
 	public provideDocumentFormattingEdits(document: vscode.TextDocument, options: vscode.FormattingOptions, token: vscode.CancellationToken): Thenable<vscode.TextEdit[]> {
 		return document.save().then(() => {
@@ -30,30 +22,8 @@ export class GoDocumentFormattingEditProvider implements vscode.DocumentFormatti
 	}
 
 	private doFormatDocument(document: vscode.TextDocument, options: vscode.FormattingOptions, token: vscode.CancellationToken): Thenable<vscode.TextEdit[]> {
-		return new Promise((resolve, reject) => {
-			var filename = document.fileName;
-
-			var formatCommandBinPath = getBinPath(this.formatCommand);
-
-			cp.execFile(formatCommandBinPath, [filename], {}, (err, stdout, stderr) => {
-				try {
-					if (err && (<any>err).code == "ENOENT") {
-						vscode.window.showInformationMessage("The '" + formatCommandBinPath + "' command is not available.  Please check your go.formatCommand user setting and ensure it is installed.");
-						return resolve(null);
-					}
-					if (err) return reject("Cannot format due to syntax errors.");
-					var text = stdout.toString();
-					// TODO: Should use `-d` option to get a diff and then compute the
-					// specific edits instead of replace whole buffer
-					var lastLine = document.lineCount;
-					var lastLineLastCol = document.lineAt(lastLine - 1).range.end.character;
-					var range = new vscode.Range(0, 0, lastLine - 1, lastLineLastCol);
-					return resolve([new vscode.TextEdit(range, text)]);
-				} catch (e) {
-					reject(e);
-				}
-			});
-		});
+		var formatter = new GoFormatter();
+		return formatter.provideDocumentFormattingEdits(document);
 	}
 
 }
